+++ app/src/main/java/com/example/playlistmaker/playlist_recycler/PlaylistAdapter.kt
@@ -1,0 +1,24 @@
+package com.example.playlistmaker.playlist_recycler
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.example.playlistmaker.R
+import com.example.playlistmaker.entities.Track
+
+class PlaylistAdapter : RecyclerView.Adapter<PlaylistViewHolder>(){
+
+    var trackList = ArrayList<Track>()
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PlaylistViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.track_view, parent, false)
+        return PlaylistViewHolder(view)
+    }
+
+    override fun getItemCount(): Int = trackList.size
+
+    override fun onBindViewHolder(holder: PlaylistViewHolder, position: Int) {
+        holder.bind(trackList[position])
+    }
+
+}
