--- app/src/main/res/drawable/album_placeholder.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="312dp"
-    android:height="312dp"
-    android:viewportWidth="312"
-    android:viewportHeight="312">
-  <path
-      android:pathData="M59.67,89.57C68.77,80.34 77.87,72.02 81.9,73.71C88.4,76.31 81.9,87.1 78,93.47C74.75,98.93 40.82,144.04 40.82,175.5C40.82,192.14 47.06,205.92 58.24,214.24C67.99,221.52 80.86,223.73 92.56,220.22C106.47,216.19 117.91,202.02 132.34,184.21C148.07,164.84 169.13,139.49 185.38,139.49C206.57,139.49 206.83,152.62 208.26,162.76C159.12,171.08 138.32,210.47 138.32,232.57C138.32,254.67 157.04,272.74 180.05,272.74C201.24,272.74 235.82,255.45 241.02,193.44H273V160.94H240.89C238.94,139.49 226.72,106.34 188.5,106.34C159.25,106.34 134.16,131.17 124.28,143.26C116.74,152.75 97.5,175.5 94.51,178.62C91.26,182.52 85.67,189.54 80.08,189.54C74.23,189.54 70.72,178.75 75.4,164.58C79.95,150.41 93.6,127.4 99.45,118.82C109.59,104 116.35,93.86 116.35,76.18C116.35,47.97 95.03,39 83.72,39C66.56,39 51.61,52 48.36,55.25C43.68,59.93 39.78,63.83 36.92,67.34L59.67,89.57ZM180.44,241.15C176.41,241.15 170.82,237.77 170.82,231.79C170.82,223.99 180.31,203.19 208.13,195.91C204.23,230.88 189.54,241.15 180.44,241.15Z"
-      android:fillColor="#AEAFB4"/>
-</vector>
