--- app/src/main/res/drawable/dark_play_btn.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="100dp"
-    android:height="100dp"
-    android:viewportWidth="100"
-    android:viewportHeight="100">
-  <path
-      android:pathData="M50,8.33C27,8.33 8.33,27 8.33,50C8.33,73 27,91.67 50,91.67C73,91.67 91.67,73 91.67,50C91.67,27 73,8.33 50,8.33ZM41.67,68.75V31.25L66.67,50L41.67,68.75Z"
-      android:fillColor="#ffffff"/>
-</vector>
