--- app/src/main/res/drawable/like_light.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="51dp"
-    android:height="51dp"
-    android:viewportWidth="51"
-    android:viewportHeight="51">
-  <path
-      android:pathData="M25.5,0L25.5,0A25.5,25.5 0,0 1,51 25.5L51,25.5A25.5,25.5 0,0 1,25.5 51L25.5,51A25.5,25.5 0,0 1,0 25.5L0,25.5A25.5,25.5 0,0 1,25.5 0z"
-      android:fillColor="#1A1B22"
-      android:fillAlpha="0.25"/>
-  <path
-      android:pathData="M31.125,15C28.95,15 26.862,16.015 25.5,17.62C24.138,16.015 22.05,15 19.875,15C16.025,15 13,18.033 13,21.894C13,26.632 17.25,30.492 23.688,36.358L25.5,38L27.313,36.346C33.75,30.492 38,26.632 38,21.894C38,18.033 34.975,15 31.125,15ZM25.625,34.491L25.5,34.616L25.375,34.491C19.425,29.088 15.5,25.516 15.5,21.894C15.5,19.387 17.375,17.507 19.875,17.507C21.8,17.507 23.675,18.748 24.337,20.465H26.675C27.325,18.748 29.2,17.507 31.125,17.507C33.625,17.507 35.5,19.387 35.5,21.894C35.5,25.516 31.575,29.088 25.625,34.491Z"
-      android:fillColor="#ffffff"/>
-</vector>
