+++ app/src/main/res/layout/activity_settings.xml
@@ -29,12 +29,11 @@
         <TextView style="@style/SettingsText"
             android:text="@string/dark_theme" />
 
-        <com.google.android.material.switchmaterial.SwitchMaterial
-            android:id="@+id/themeSwitch"
+        <Switch
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="end|center"
-            />
+            android:layout_gravity="end"
+            tools:ignore="UseSwitchCompatOrMaterialXml" />
 
     </FrameLayout>
