+++ app/src/main/res/layout/track_view.xml
@@ -3,7 +3,6 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="horizontal"
-    android:foreground="?attr/selectableItemBackground"
     android:paddingTop="@dimen/edittext_icon_padding"
     android:paddingBottom="@dimen/edittext_icon_padding">
