+++ app/src/main/res/values-night/styles.xml
@@ -9,75 +9,6 @@
         <item name="android:background">@color/black</item>
     </style>
 
-    <style name="PlayerBackBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/dark_back_btn</item>
-    </style>
-
-    <style name="PlayerPlaylistBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/playlist_dark</item>
-    </style>
-
-    <style name="PlayerLikeBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/like_dark</item>
-    </style>
-
-    <style name="PlayerPlayBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/dark_play_btn</item>
-    </style>
-
-    <style name="PlayerTrackName">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:layout_marginTop">@dimen/dp24</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textSize">@dimen/sp24</item>
-        <item name="android:textColor">@color/white</item>
-    </style>
-
-    <style name="PlayerTrackArtist">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:layout_marginTop">@dimen/dp12</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textSize">@dimen/sp16</item>
-        <item name="android:textColor">@color/white</item>
-    </style>
-
-    <style name="PlayerTrackTimer">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingBottom">@dimen/dp28</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textColor">@color/white</item>
-    </style>
-
-    <style name="PlayerLeftText">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingTop">@dimen/dp8</item>
-        <item name="android:paddingBottom">@dimen/dp8</item>
-        <item name="android:fontFamily">@font/ys_display_regular</item>
-        <item name="android:textColor">@color/grey</item>
-    </style>
-
-    <style name="PlayerRightText">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingTop">@dimen/dp8</item>
-        <item name="android:paddingBottom">@dimen/dp8</item>
-        <item name="android:fontFamily">@font/ys_display_regular</item>
-        <item name="android:textColor">@color/white</item>
-        <item name="android:layout_gravity">end</item>
-    </style>
-
     <style name="TrackButton">
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
@@ -120,12 +51,11 @@
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:gravity">center</item>
-        <item name="android:layout_gravity">center</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textSize">@dimen/sp20</item>
+        <item name="android:textSize">@dimen/l_text</item>
         <item name="android:textColor">@color/white</item>
-        <item name="android:paddingTop">@dimen/dp16</item>
-        <item name="android:paddingBottom">@dimen/dp16</item>
+        <item name="android:paddingTop">@dimen/l_size</item>
+        <item name="android:paddingBottom">@dimen/l_size</item>
     </style>
 
     <style name="TrackName">
@@ -203,20 +133,20 @@
         <item name="android:textColor">@color/white</item>
         <item name="android:textSize">@dimen/medium_text_size</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
-        <item name="android:paddingTop">@dimen/settings_padding</item>
-        <item name="android:paddingBottom">@dimen/settings_padding</item>
     </style>
 
     <style name="SettingsFrame">
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
+        <item name="android:paddingTop">@dimen/settings_padding</item>
         <item name="android:foreground">?attr/selectableItemBackground</item>
+        <item name="android:paddingBottom">@dimen/settings_padding</item>
     </style>
 
     <style name="SettingsIcon">
         <item name="android:layout_width">@dimen/settings_icon_size</item>
         <item name="android:layout_height">@dimen/settings_icon_size</item>
-        <item name="android:layout_gravity">end|center</item>
+        <item name="android:layout_gravity">end</item>
         <item name="tint">@color/white</item>
     </style>
