+++ app/src/main/res/values/styles.xml
@@ -13,7 +13,7 @@
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:layout_gravity">center|end</item>
-        <item name="android:tint">@color/grey</item>
+        <item name="android:tint">@color/searchbar_text_light</item>
     </style>
 
     <style name="NoServiceButton">
@@ -50,12 +50,11 @@
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:gravity">center</item>
-        <item name="android:layout_gravity">center</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textSize">@dimen/sp20</item>
+        <item name="android:textSize">@dimen/l_text</item>
         <item name="android:textColor">@color/black</item>
-        <item name="android:paddingTop">@dimen/dp16</item>
-        <item name="android:paddingBottom">@dimen/dp16</item>
+        <item name="android:paddingTop">@dimen/l_size</item>
+        <item name="android:paddingBottom">@dimen/l_size</item>
     </style>
 
     <style name="TrackName">
@@ -67,89 +66,20 @@
         <item name="android:layout_gravity">left</item>
     </style>
 
-    <style name="PlayerBackBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/light_back_btn</item>
-    </style>
-
-    <style name="PlayerPlaylistBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/playlist_light</item>
-    </style>
-
-    <style name="PlayerLikeBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/like_light</item>
-    </style>
-
-    <style name="PlayerPlayBtn">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:src">@drawable/light_play_btn</item>
-    </style>
-
-    <style name="PlayerTrackName">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:layout_marginTop">@dimen/dp24</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textSize">@dimen/sp24</item>
-        <item name="android:textColor">@color/black</item>
-    </style>
-
-    <style name="PlayerTrackArtist">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:layout_marginTop">@dimen/dp12</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textSize">@dimen/sp16</item>
-        <item name="android:textColor">@color/black</item>
-    </style>
-
-    <style name="PlayerTrackTimer">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingBottom">@dimen/dp28</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textColor">@color/black</item>
-    </style>
-
-    <style name="PlayerLeftText">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingTop">@dimen/dp8</item>
-        <item name="android:paddingBottom">@dimen/dp8</item>
-        <item name="android:fontFamily">@font/ys_display_regular</item>
-        <item name="android:textColor">@color/grey</item>
-    </style>
-
-    <style name="PlayerRightText">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingTop">@dimen/dp8</item>
-        <item name="android:paddingBottom">@dimen/dp8</item>
-        <item name="android:fontFamily">@font/ys_display_regular</item>
-        <item name="android:textColor">@color/black</item>
-        <item name="android:layout_gravity">end</item>
-    </style>
-
     <style name="SeparateDot">
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:layout_gravity">center</item>
         <item name="android:paddingStart">@dimen/separate_dot</item>
         <item name="android:paddingEnd">@dimen/separate_dot</item>
-        <item name="android:tint">@color/grey</item>
+        <item name="android:tint">@color/searchbar_text_light</item>
     </style>
 
     <style name="TrackSubtext">
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:textSize">@dimen/track_subtext</item>
-        <item name="android:textColor">@color/grey</item>
+        <item name="android:textColor">@color/searchbar_text_light</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
         <item name="android:layout_gravity">center</item>
     </style>
@@ -166,10 +96,10 @@
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:textSize">@dimen/medium_text_size</item>
-        <item name="android:textColorHint">@color/grey</item>
+        <item name="android:textColorHint">@color/searchbar_text_light</item>
         <item name="android:background">@drawable/rounded_edittext</item>
         <item name="android:backgroundTint">@color/searchbar</item>
-        <item name="android:drawableTint">@color/grey</item>
+        <item name="android:drawableTint">@color/searchbar_text_light</item>
         <item name="android:drawablePadding">@dimen/edittext_icon_padding</item>
         <item name="android:padding">@dimen/edittext_padding</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
@@ -200,20 +130,20 @@
         <item name="android:textColor">@color/black</item>
         <item name="android:textSize">@dimen/medium_text_size</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
-        <item name="android:paddingTop">@dimen/settings_padding</item>
-        <item name="android:paddingBottom">@dimen/settings_padding</item>
     </style>
 
     <style name="SettingsFrame">
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
+        <item name="android:paddingTop">@dimen/settings_padding</item>
         <item name="android:foreground">?attr/selectableItemBackground</item>
+        <item name="android:paddingBottom">@dimen/settings_padding</item>
     </style>
 
     <style name="SettingsIcon">
         <item name="android:layout_width">@dimen/settings_icon_size</item>
         <item name="android:layout_height">@dimen/settings_icon_size</item>
-        <item name="android:layout_gravity">end|center</item>
+        <item name="android:layout_gravity">end</item>
     </style>
 
     <style name="MainScreenButton" parent="Widget.MaterialComponents.Button.UnelevatedButton">
